utilisation de long pour la réponse des PIDs

+++ PID.c
@@ -1,10 +1,10 @@
 #include "PID.h"
 #include "reglages.h"
 
-int PID_lineique(int erreur, int erreur_preced, int erreur_sum)
+long int PID_lineique(int erreur, int erreur_preced, int erreur_sum)
 {
 //erreur_sum intégration par somme ou méthode des trapèzes ?
-	int commande;
+	long int commande;
 
 	commande= KP_DELTA*erreur; //composante proportionelle
 	commande+= KI_DELTA*erreur_sum; //composante intégrale
@@ -13,10 +13,10 @@ int PID_lineique(int erreur, int erreur_preced, int erreur_sum)
 	return commande;
 }
 
-int PID_angulaire(int erreur, int erreur_preced, int erreur_sum)
+long int PID_angulaire(int erreur, int erreur_preced, int erreur_sum)
 {
 //erreur_sum intégration par somme ou méthode des trapèzes ?
-	int commande;
+	long int commande;
 
 	commande= KP_ALPHA*erreur; //composante proportionelle
 	commande+= KI_ALPHA*erreur_sum; //composante intégrale
