+++ PID.h
@@ -1,2 +1,2 @@
-int PID_lineique(int erreur, int erreur_preced, int erreur_sum);
-int PID_angulaire(int erreur, int erreur_preced, int erreur_sum);
+long int PID_lineique(int erreur, int erreur_preced, int erreur_sum);
+long int PID_angulaire(int erreur, int erreur_preced, int erreur_sum);
